camera locked into window height

+++ src/player/camera.rs
@@ -23,15 +23,22 @@ fn setup_system(commands: &mut Commands) {
 }
 
 fn camera_follow_system(
-    mut query1: Query<&mut Transform, With<Camera>>,
-    query2: Query<&Transform, With<Player>>,
+    windows: Res<Windows>,
+    mut camera_query: Query<&mut Transform, With<Camera>>,
+    player_query: Query<&Transform, With<Player>>,
 ) {
     let smooth_speed: f32 = 0.125;
     let offset = Vec2::new(0., 2.);
+    let window = windows.get_primary().unwrap();
+    let half_height = window.height() as f32 * 0.5;
 
-    for mut transform in query1.iter_mut() {
-        for target_transform in query2.iter() {
+    for mut transform in camera_query.iter_mut() {
+        for target_transform in player_query.iter() {
             let desired_position = target_transform.translation.truncate() + offset;
+            if desired_position.y >= half_height || desired_position.y <= -(half_height) {
+                return;// we stop camera if player goes out of bounds
+            }
+
             let smoothed_position = transform
                 .translation
                 .truncate()
